% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper_vis.R
\name{auto_set_colorcode}
\alias{auto_set_colorcode}
\title{Create color codes for groups.}
\usage{
auto_set_colorcode(groups, palette = "Set1")
}
\arguments{
\item{groups}{A vector giving the names of groups. Duplicate group names
will be disregarded.}

\item{palette}{A string giving the name of palette provided in
\code{\link[RColorBrewer:brewer.pal]{RColorBrewer}}.}
}
\value{
A dataframe that contains color codes for samples from different
  groups. The first column consists of names of the groups and the second
  column contains the coresponding color codes.
}
\description{
\code{auto_set_colorcode} returns a dataframe with group names and color
codes.
}
\examples{
auto_set_colorcode(c('a','b','c'))

}
