% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_season.R
\name{check_season}
\alias{check_season}
\title{Check seasonality.}
\usage{
check_season(x, season_from, season_to)
}
\arguments{
\item{x}{A ts, zoo or xts serie}

\item{season_from}{String date from initial season. i.e. '2017-06-01'}

\item{season_to}{String date to end season. i.e. '2017-08-30'}
}
\value{
ts, zoo or xts
}
\description{
Given a historical series and seasonality it returns a column
indicating whether or not each record is in its season.
}
\details{
\code{check_season} returns a ts, zoo or xts indicating seasonality.
}
\examples{
\dontrun{check_season('SPY', season_from = '2017-01-01', season_to = '2017-03-30')}
}
\author{
Mario Pisa
}
\seealso{
\url{}
}

