% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper_vis.R
\name{color_map}
\alias{color_map}
\title{Assign color codes to samples}
\usage{
color_map(samples_group, color_code)
}
\arguments{
\item{samples_group}{A vector of group names of samples}

\item{color_code}{A color code dataframe}
}
\value{
A vector of color codes matching the group names
}
\description{
\code{color_map} assigns color codes to samples based on the given color code.
}
\examples{
temp_code = auto_set_colorcode(c('a','b','c'))
temp_groups = c('a','a','b','c')
color_map(temp_groups, temp_code)

}
