% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_mixer}
\alias{color_mixer}
\title{Color mixer}
\usage{
color_mixer(col_vec, weight = NULL, na.rm = TRUE)
}
\arguments{
\item{weight}{vector of weight of each color in calculating the average}

\item{colors}{vector of Hex colors}
}
\description{
Color mixer
}
\keyword{internal}
