% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_utils.R
\name{compute_vif}
\alias{compute_vif}
\title{Variance Inflation Factors}
\usage{
compute_vif(data)
}
\arguments{
\item{data}{Data frame or matrix of predictors.}
}
\value{
Named numeric vector of VIF values.
}
\description{
Computes the variance inflation factor for each predictor column, following the
helper used in Assignment 7.
}
\examples{
data("Boston", package = "MASS")
compute_vif(subset(Boston, select = -crim))

}
