% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_coordinate}
\alias{filter_coordinate}
\title{Coordinate projection filter}
\usage{
filter_coordinate(dat, k, ...)
}
\arguments{
\item{dat}{A numeric dataset matrix, rach row represents a data point and
each column represents a predictive variable.}

\item{k}{A scalar or vector deciding to which coordinate(s) the data matrix is
projected.}

\item{...}{Further arguments.}
}
\value{
A matrix object of filter values.
}
\description{
\code{filter_coordinate} computes the Coordinate projection filter.
}
\details{
Coordinate projection filter is defined as \eqn{f(x_i;k) = x_{ik}}.
}
\examples{
tp_data = chicken_generator(1)
filter_coordinate(tp_data[,-1],1)

}
