% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_dtm}
\alias{filter_dtm}
\title{Distance to measure filter}
\usage{
filter_dtm(dist, k, p = 2, ...)
}
\arguments{
\item{dist}{The distance matrix.}

\item{k}{A numeric scalar deciding the number of neighbors to be considered}

\item{p}{An optional numeric scalar denoting the exponent.}

\item{...}{Further arguments.}
}
\value{
A matrix object of filter values.
}
\description{
\code{filter_dtm} computes the Distance to measure filter.
}
\details{
Distance to measure filter function is defined as \eqn{f(x_i;k) =
\frac{1}{k}\sum_{j=2}^{k+1} d^p(x_i,x_{(j)})^{1/p}}.
}
\examples{
tp_data = chicken_generator(1)
tp_dist = dist(tp_data[,-1])
filter_dtm(dist = tp_dist, k=3)
}
