% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_eccen}
\alias{filter_eccen}
\title{Eccentricity filter}
\usage{
filter_eccen(dist, p, ...)
}
\arguments{
\item{dist}{The distance matrix.}

\item{p}{A numeric scalar denoting the exponent.}

\item{...}{Further arguments.}
}
\value{
A matrix object of filter values.
}
\description{
\code{filter_eccen} computes the Eccentricity filter.
}
\details{
Eccentricity filter function is defined as \eqn{f(x_i;p) =
(\frac{1}{n}\sum_{j=1}^n d^p(x_i,x_j))^{1/p}}.
}
\examples{
tp_data = chicken_generator(1)
tp_dist = dist(tp_data[,-1])
filter_eccen(dist = tp_dist, p = 2)

}
