% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{filter_evaluate}
\alias{filter_evaluate}
\title{Evaluate filter functions}
\usage{
filter_evaluate(..., dat, group_ind, num_intervals, percent_overlap,
  arg_mapper = list())
}
\arguments{
\item{...}{Filter objects. The classes of the objects should be \code{filter}
and include an attribute \code{filter} which is the name of the
corresponding filter function.}

\item{dat}{Matrix or dataset where rows are data points and columns are
predictive variables.}

\item{group_ind}{A vector of group names each of the samples belongs to.}

\item{num_intervals}{A length m vector of positive integers.}

\item{percent_overlap}{A length m vector of numbers between 0 and 100 specifying how much adjacent intervals should overlap.}

\item{arg_mapper}{A list for additional arguments for
\code{\link{mapper.kmeans}}.}
}
\value{
A data.frame of Shannon indices and spread measures under given
  filter functions. The first column contents names of filter functions, and
  the second and third columns are the shannon indices and spread measures,
  respectively.
}
\description{
\code{filter_evaluate} evaluates filter function with provided filter
vectors.
}
\examples{
tp_data = chicken_generator(1)
tp_dist = dist(tp_data[,-1])
a = filter_eccen(dist = tp_dist, p = 2)
b = filter_coordinate(tp_data[,-1], 2)
c = filter_gaussian(dist=tp_dist, sigma=1)
filter_evaluate(a,b,c,
                dat = tp_data, group_ind = tp_data$Group,
                num_intervals = 10, percent_overlap = 70)

# Add additional arguments (NOT RUN)
if(FALSE) {
  filter_evaluate(a,b,c,
                  dat = tp_data, group_ind = tp_data$Group,
                  num_intervals = 10, percent_overlap = 70,
                  arg_mapper = list(n_class = 1))
}

}
