% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_gaussian}
\alias{filter_gaussian}
\title{Gaussian density filter}
\usage{
filter_gaussian(dist, sigma = 1, ...)
}
\arguments{
\item{dist}{The distance matrix.}

\item{sigma}{A scalar controling the sensitivity of the gaussian kernel.}

\item{...}{Further arguments.}
}
\value{
A matrix object of filter values.
}
\description{
\code{filter_gaussian} computes the Gaussian density filter.
}
\details{
The Gaussian density filter is defined as \eqn{f(x_i;\sigma) = C\sum_{j=1}^n
exp(-\frac{\|x_i-x_j\|^2}{2\sigma^2})}, where \eqn{C} is the normalizing
constant and \eqn{\sigma} is the scalar controling the sensitivity of the
gaussian kernel.
}
\examples{
tp_data = chicken_generator(1)
tp_dist = dist(tp_data[,-1])
filter_gaussian(dist=tp_dist, sigma=1)

}
