% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_Linf}
\alias{filter_Linf}
\title{L-infinity filter}
\usage{
filter_Linf(dist, ...)
}
\arguments{
\item{dist}{The distance matrix.}

\item{...}{Optional arguments to \code{\link[base]{max}}.}
}
\value{
A matrix object of filter values.
}
\description{
\code{filter_Linf} computes the L-infinity filter.
}
\details{
The L-infinity filter function is defined as \eqn{f(x_i) = max_{j} d(x_i,x_j)} where
\eqn{x_i,x_j} are some data points.
}
\examples{
tp_data = chicken_generator(1)
tp_dist = dist(tp_data[,-1])
filter_Linf(dist=tp_dist)

}
