% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_name}
\alias{filter_name}
\title{Which filter function produces the filter?}
\usage{
filter_name(filter_vec)
}
\arguments{
\item{filter_vec}{The filter vector.}
}
\value{
The name of the filter.
}
\description{
\code{filter_name} returns the name of the filter function that produces the
provided filter.
}
\examples{
tp_data = chicken_generator(1)
tp_dist = dist(tp_data[,-1])
res = filter_eccen(dist = tp_dist, p = 2)
filter_name(res)
## "Eccentricity"
}
