% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_pca}
\alias{filter_pca}
\title{PCA filter}
\usage{
filter_pca(dat, k = 1, ...)
}
\arguments{
\item{dat}{A numeric dataset matrix, rach row represents a data point and
each column represents a predictive variable.}

\item{k}{A scaler deciding the number of principle components to be returned}

\item{...}{Optional arguments to \code{\link[stats]{cov}}.}
}
\value{
A matrix object of filter values.
}
\description{
\code{filter_pca} returns the first k principle components as filter values.
}
\details{
The PCA filter function is defined as \eqn{f(x_i) = x_i^T\phi_{(1:k)}}, where
\eqn{\phi_{(1:k)}} is the matrix of \eqn{k} eigenvectors associated with the
\eqn{k} largest eigenvalues of \eqn{cov(X)}, \eqn{x_i} is some data point and
\eqn{X} is the data matrix.
}
\examples{
tp_data = chicken_generator(1)
filter_pca(dat=tp_data[,-1])

}
