% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_ref}
\alias{filter_ref}
\title{Reference distance filter}
\usage{
filter_ref(dist, groups_ind, ref, ...)
}
\arguments{
\item{dist}{The distance matrix.}

\item{groups_ind}{A vector of group names each of the samples belongs to.}

\item{ref}{A string object specifying the name of the reference group.}

\item{...}{Additional arguments.}
}
\value{
A matrix object of filter values.
}
\description{
\code{filter_ref} computes the reference distance filter.
}
\details{
Reference distance filter function is defined as \eqn{f(x_i;y,g) =
1-\textrm{median}_{y_j=g} d(x_i,x_j)}, where
}
\examples{
tp_data = chicken_generator(1)
tp_dist = dist(tp_data[,-1])
filter_ref(dist=tp_dist, groups_ind=tp_data$Group, ref=4)

}
