% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staFunctions.R
\name{getMaster}
\alias{getMaster}
\title{Get a \code{RasterLayer} with no missing values from a \code{Raster*} object}
\usage{
getMaster(x)
}
\arguments{
\item{x}{\code{Raster*} object}
}
\value{
\code{RasterLayer}
}
\description{
The term \emph{master} refers to a raster layer whose extent and coordinate reference system
are used as a reference to rasterize further objects, e.g. matrices. To rasterize, \emph{master} must be free of missing values.
}
\seealso{
\code{\link[geoTS]{matrixToRaster}}
}
