% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InverseCDF.R
\name{Inverse.CDF}
\alias{Inverse.CDF}
\title{Inverse CDF Sampling}
\usage{
Inverse.CDF(iCDF, n, ...)
}
\arguments{
\item{iCDF}{Function that maps probabilities in (0, 1) to quantiles.}

\item{n}{Integer number of samples to generate.}

\item{...}{Additional arguments passed to `iCDF`.}
}
\value{
Numeric vector of simulated values.
}
\description{
Generates random draws by transforming uniform samples through a supplied inverse
CDF, as demonstrated in Assignment 6 notes.
}
\examples{
inv_exp <- function(u, rate) -log(1 - u) / rate
draws <- Inverse.CDF(inv_exp, n = 1000, rate = 2)

}
