% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper_vis.R
\name{legend_node}
\alias{legend_node}
\title{Add legends to the generated graph}
\usage{
legend_node(stats_sum = NULL, color_code)
}
\arguments{
\item{stats_sum}{Results from stat_summery}

\item{color_code}{Color code dataframe}
}
\value{
Dataframe of legends for the generated graph
}
\description{
Add legends to the generated graph
}
