% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_network_h5}
\alias{load_network_h5}
\title{Load a h5 file}
\usage{
load_network_h5(file = "network.h5")
}
\arguments{
\item{file}{The filename (character) of the file in which the dataset will be
located.}
}
\value{

}
\description{
Load a h5 file from save_network_h5.
}
