% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{mapper.sta}
\alias{mapper.sta}
\title{Mapper function with multiple cluster methods}
\usage{
mapper.sta(
  dat,
  filter_values,
  num_intervals,
  percent_overlap,
  dist_method = "euclidean",
  cluster_method = "kmeans",
  NbClust_cluster_method = "kmeans",
  num_bins_when_clustering = 10,
  cluster_index = "all",
  n_class = 0,
  eps = 0.15,
  minPts = 5,
  permute_interval_level = FALSE,
  ...
)
}
\arguments{
\item{dat}{Matrix or dataset where rows are data points and columns are
predictive variables.}

\item{filter_values}{A n x m data frame of real numbers returned by the
filter functions.}

\item{dist_method}{The distance measure to be used to compute the
dissimilarity matrix. By default, distance="euclidean". It must be one of
This must be one of: "euclidean", "maximum", "manhattan", "canberra",
"binary", "minkowski" or "NULL". Details can be found in
\code{\link[NbClust]{NbClust}}.}

\item{cluster_method}{Clustering method. This should be one of:
"hierarchical", "kmeans", "dbscan", "hdbscan".}

\item{NbClust_cluster_method}{The cluster analysis method to be used. This
should be one of: "ward.D", "ward.D2", "single", "complete", "average",
"mcquitty", "median", "centroid", "kmeans".Details can be found in
\code{\link[NbClust]{NbClust}}.}

\item{num_bins_when_clustering}{For hierachical clustering. A positive
integer that controls whether points in the same level set end up in the
same cluster.}

\item{cluster_index}{The index to be calculated. This should be one of :
"kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew",
"friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2",
"beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain",
"gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw",
"all" (all indices except GAP, Gamma, Gplus and Tau), "alllong" (all
indices with Gap, Gamma, Gplus and Tau included). Details can be found in
\code{\link[NbClust]{NbClust}}.}

\item{n_class}{number of clusters for k means. By default, n_class=0. If
n_class>0, this function will instead call \code{\link[stats]{kmeans}} and
pass \code{n_class} to argument \code{centers} of
\code{\link[stats]{kmeans}}.}

\item{eps}{for DBSCAN, size of the epsilon neighborhood}

\item{minPts}{for DBSCAN and HDBSCAN, number of minimum points in the eps
region for core points. Default is 2 points}

\item{permute_interval_level}{boolean. True if samples within each interval
are to be permuted}

\item{...}{Further arguments for \code{\link[NbClust]{NbClust}} or
\code{\link[stats]{kmeans}} or \code{\link[stats]{hclust}} or
\code{\link[dbscan]{dbscan}} or \code{\link[dbscan]{hdbscan}}}
}
\value{
An object of class \code{TDAmapper} which is a list of items named
  \code{adjacency} (adjacency matrix for the edges), \code{num_vertices}
  (integer number of vertices), \code{level_of_vertex} (vector with
  \code{level_of_vertex[i]} = index of the level set for vertex i),
  \code{points_in_vertex} (list with \code{points_in_vertex[[i]]} = vector of
  indices of points in vertex i), \code{points_in_level} (list with
  \code{points_in_level[[i]]} = vector of indices of points in level set i,
  and \code{vertices_in_level} (list with \code{vertices_in_level[[i]]} =
  vector of indices of vertices in level set i.
}
\description{
This function is adopted from \code{mapper} function of \code{TDAmapper} with
different clustering methods (mainly k-means).
}
\details{
This function is adopted from \code{mapper} function of \code{TDAmapper} by
replacing its cluster method with the cluster function
\code{\link[NbClust]{NbClust}} from R package \code{NbClust}.

The advantage of \code{NbClust} is that it provides users with 8 different
cluster methods, 6 different distance measures and 30 indices for determining
the number of clusters. This allows users to select the best clustering
scheme from the different results obtained by varying all combinations of
number of clusters, distance measures, and clustering methods. Details of the
distance measures, clustering methods and cluster indices can be found in
\code{\link[NbClust]{NbClust}}.
}
\examples{
tp_data = chicken_generator(1)
tp_data_mapper = mapper.sta(dat = tp_data[,2:4],
                               filter_values = tp_data$Y,
                               num_intervals = 10,
                               percent_overlap = 70)

}
\references{
Malika Charrad, Nadia Ghazzali, Veronique Boiteau, Azam Niknafs
  (2014). NbClust: An R Package for Determining the Relevant Number of
  Clusters in a Data Set. Journal of Statistical Software, 61(6), 1-36. URL
  http://www.jstatsoft.org/v61/i06/.
}
