% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mapper_shannon_index}
\alias{mapper_shannon_index}
\title{Shannon index for Mapper}
\usage{
mapper_shannon_index(obj_mapper, group_ind)
}
\arguments{
\item{obj_mapper}{An object of class \code{TDAmapper}.}

\item{group_ind}{A vector of group names each of the samples belongs to.}
}
\value{
A list object with two objects. The first object is the weighted
  average Shannon index for the whole graph. The second object is the matrix
  of Shannon indices for each of the nodes in the graph, where the first
  column is the vector of indices and the second column is the vector of node
  sizes.
}
\description{
\code{mapper_shannon_index} computes Shannon indices for nodes of a given
graph from Mapper.
}
\details{
how it is computed
}
\examples{
library(TDAmapper)
tp_data = chicken_generator(1)
tp_data_mapper = mapper.kmeans(dat = tp_data[,2:4],
                               filter_values = tp_data$Y,
                               num_intervals = 10,
                               percent_overlap = 70)
tp_shannon = mapper_shannon_index(tp_data_mapper, tp_data$Group)
tp_shannon

}
