% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapperEdges}
\alias{mapperEdges}
\title{mapperEdges function}
\usage{
mapperEdges(m)
}
\arguments{
\item{m}{An object of class TDAmapper that is the output of the mapper function.}
}
\value{
A data frame describing the edges in the graph of the mapper output.
}
\description{
The input to this function is a TDAmapper class object and the output
is a data frame of edges that can be used as input to the networkD3
plot utility.
}
\references{
\url{https://github.com/paultpearson/TDAmapper}
}
\seealso{
\code{\link{mapperVertices}}
}
\author{
Paul Pearson, \email{pearsonp@hope.edu}
}
\keyword{mapperEdges}
