% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapperVertices}
\alias{mapperVertices}
\title{mapperVertices function}
\usage{
mapperVertices(m, pt_labels)
}
\arguments{
\item{m}{An object of class TDAmapper that is the output of the mapper
function.}
}
\value{
A data frame describing the vertices in the graph of the mapper
output and the point labels that will be displayed when the mouse
hovers over a vertex in the graph.
}
\description{
The input to this function is a TDAmapper class object and the output
is a data frame of vertices that can be used as input to the networkD3
plot utility.
}
\references{
\url{https://github.com/paultpearson/TDAmapper}
}
\seealso{
\code{\link{mapperEdges}}
}
\author{
Paul Pearson, \email{pearsonp@hope.edu}
}
\keyword{internal}
\keyword{mapperVertices}
