% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marismas}
\alias{marismas}
\title{10-day Composite NDMI Time Series}
\format{
An object of class \code{"numeric"}.
}
\usage{
data(marismas)
}
\description{
A \code{numeric} vector of length 684 containing 10-day composite values of the Normalized Difference Moisture Index (NDMI) from 2000 to 2018.
NDMI = (NIR-MIR)/(NIR+MIR) where NIR and MIR are the Near Infrared and Mid-Infrared bands of the MCD43A4 MODIS 
product, respectively. This \code{numeric} vector was taken from a \code{RasterStack} covering the Natural
Protected Area \emph{Reserva de la Biosfera Marismas Nacionales} at Nayarit, Mexico.
}
\keyword{datasets}
