% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_remover}
\alias{null_remover}
\title{Remove NULL nodes in the graph}
\usage{
null_remover(obj_mapper)
}
\arguments{
\item{obj_mapper}{The Mapper object}
}
\value{
A Mapper object
}
\description{
Remove NULL nodes in the graph
}
\keyword{internal}
