% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta.R
\name{plot.staMatrix}
\alias{plot.staMatrix}
\title{Plot method for \code{sta} function}
\usage{
\method{plot}{staMatrix}(x, ..., significance, master)
}
\arguments{
\item{x}{an object of class "staMatrix"}

\item{...}{additional plot parameters}

\item{significance}{numeric indicating significance of each shape parameter trend}

\item{master}{character with full path to a \emph{.tif} file which is used to transfer STA
output to a raster file.}
}
\description{
This function returns a \code{\link[mapview]{mapview}} object
}
\seealso{
\code{\link[sta]{sta}}
}
