% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta.R
\name{plot.staNumeric}
\alias{plot.staNumeric}
\title{Plot method for \code{sta} function}
\usage{
\method{plot}{staNumeric}(x, significance = NULL, ...)
}
\arguments{
\item{x}{an object of class "staNumeric"}

\item{significance}{numeric indicating significance of each shape parameter trend}

\item{...}{additional plot parameters}
}
\description{
This function returns a plot
}
\seealso{
\code{\link{sta}}
}
