% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{print.filter}
\alias{print.filter}
\title{Print filter function name}
\usage{
\method{print}{filter}(x)
}
\arguments{
\item{x}{a filter object}
}
\value{
the name of the filter function producing the filter
}
\description{
Print filter function name
}
\seealso{
\code{\link{filter_pca}}
}
