% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{print.Pareto_frontier}
\alias{print.Pareto_frontier}
\title{Print Pareto frontier}
\usage{
\method{print}{Pareto_frontier}(res, ...)
}
\arguments{
\item{res}{The Pareto_frontier object}
}
\value{
Summary of the Pareto_frontier object
}
\description{
Print Pareto frontier
}
\seealso{
\code{\link{pareto_opt}}
}
