% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper_vis.R
\name{read_color_code}
\alias{read_color_code}
\title{Read color codes from a given file}
\usage{
read_color_code(file, header = FALSE, sep = "\\t")
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
    Each row of the table appears as one line of the file.  If it does
    not contain an \emph{absolute} path, the file name is
    \emph{relative} to the current working directory,
    \code{\link{getwd}()}. Tilde-expansion is performed where supported.
    This can be a compressed file (see \code{\link{file}}).

    Alternatively, \code{file} can be a readable text-mode
    \link{connection} (which will be opened for reading if
    necessary, and if so \code{\link{close}}d (and hence destroyed) at
    the end of the function call).  (If \code{\link{stdin}()} is used,
    the prompts for lines may be somewhat confusing.  Terminate input
    with a blank line or an EOF signal, \code{Ctrl-D} on Unix and
    \code{Ctrl-Z} on Windows.  Any pushback on \code{stdin()} will be
    cleared before return.)

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link{url}}.)
  }

\item{header}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}

\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}
}
\value{
A dataframe that contains color codes for samples from different
  groups. The first column consists of names of the groups and the second
  column contains the coresponding color codes.
}
\description{
\code{read_color_code} reads a file with given group names and color codes
and returns a dataframe with group names and color codes that can be
accepted by visualization functions.
}
\details{
The given files should contain two columns. The first column should consist
of the names of the labels, and the second column should contain standard
Hex color codes.
}
\examples{
test1 <- data.frame(c('a','b','c'), c("#f97f6c", "#6dcff6", "#fdbc4b"))
tf <- tempfile()
write.table(test1, tf, row.names = FALSE, col.names = FALSE, sep = "\\t")

check_color_code(read_color_code(tf))
unlink(tf)
}
