% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{ref_dist_nodewise}
\alias{ref_dist_nodewise}
\title{Nodewise topological distance to a node}
\usage{
ref_dist_nodewise(obj_mapper, node_idx)
}
\arguments{
\item{obj_mapper}{An object of class \code{TDAmapper}.}

\item{node_idx}{The index of the selected node.}
}
\value{
A vector of distance to the selected node. Distance of outlier nodes
  will be set to \code{NA}.
}
\description{
This function calculates the nodewise topological distances to a selected
node based on the generated network.
}
