% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{ref_dist_samplewise}
\alias{ref_dist_samplewise}
\title{Average samplewise topological distance to a group of nodes}
\usage{
ref_dist_samplewise(obj_mapper, groups_ind, ref)
}
\arguments{
\item{obj_mapper}{An object of class \code{TDAmapper}.}

\item{groups_ind}{A vector of group names each of the samples belongs to.}

\item{ref}{A string object specifying the name of the reference group.}
}
\value{
A vector of the average distance to the selected group of samples.
  Distance of samples in the outlier nodes will be set to \code{NA}. Length
  of the vector will be the number of samples.
}
\description{
This function calculated the average topological distance of samples to a
geoup of samples.
}
