% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_network_h5}
\alias{save_network_h5}
\title{Save results into a h5 file}
\usage{
save_network_h5(obj_mapper, dataset = NULL, file = "network.h5")
}
\arguments{
\item{obj_mapper}{An object of class \code{TDAmapper}.}

\item{dataset}{The original dataset used to generate the network.}

\item{file}{The filename (character) of the file in which the dataset will be
located.}
}
\description{
Save results into a h5 file
}
