% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{spread_measure}
\alias{spread_measure}
\title{Spread measure function}
\usage{
spread_measure(obj_mapper, group_ind)
}
\arguments{
\item{obj_mapper}{An object of class \code{TDAmapper}.}

\item{group_ind}{A vector of group names each of the samples belongs to.}
}
\value{
A scalar which is the spread measure of the graph.
}
\description{
\code{spread_measure} measures how data points from the same groups
concentrate in the network.
}
\details{
How it is computed
}
\examples{
tp_data = chicken_generator(1)
tp_data_mapper = mapper.kmeans(dat = tp_data[,2:4],
                               filter_values = tp_data$Y,
                               num_intervals = 10,
                               percent_overlap = 70)
tp_spread = spread_measure(tp_data_mapper, tp_data$Group)
tp_spread

}
