% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta.R
\name{sta}
\alias{sta}
\title{Statistical Seasonal Trend Analysis}
\usage{
sta(data, freq, numFreq = 4, delta = 0.2, startYear = 2000,
  endYear = 2019, intraAnnualPeriod = c("wetSeason", "drySeason"),
  interAnnualPeriod, adhocPeriod = NULL, plot = F,
  significance = NULL, save = F, dirToSaveSTA = NULL,
  numCores = 20, master = NULL)
}
\arguments{
\item{data}{numeric vector, matrix or RasterStack object}

\item{freq}{integer with the number of observations per period. See \code{Details}}

\item{numFreq}{integer with the number of frequencies to employ in harmonic regression fitting. 
See \code{\link[geoTS]{haRmonics}}}

\item{delta}{numeric (positive) controlling regularization and prevent non-invertible
hat matrix in harmonic regression model. See \code{\link[geoTS]{haRmonics}}}

\item{startYear}{numeric, time series initial year}

\item{endYear}{numeric, time series last year}

\item{intraAnnualPeriod}{character indicating seasons (wet or dry) to be considered for additional 
statistical analysis. See \code{Details}}

\item{interAnnualPeriod}{numeric vector indicating the number of years to be considered in STA. For instance,
1:5, indicates that the first five years will be
utilized for STA. Similarly, c(2,6,10) indicates that the second, sixth and tenth
years will be utilized for STA. See \code{Details}}

\item{adhocPeriod}{numeric vector with the specific observations to be considered in additional
statistical analysis. See \code{Details}}

\item{plot}{logical, should STA maps be plotted, default is \code{FALSE}. See \code{Details}}

\item{significance}{numeric, significance level used in additional statistical analysis}

\item{save}{logical, should STA output be saved, default is \code{FALSE}}

\item{dirToSaveSTA}{character with full path name, where the STA maps will be saved, applicable when 
\code{save = TRUE}}

\item{numCores}{integer given the number of cores to use; pertinent when \code{data} is a
\code{RasterStack} or a \code{matrix}}

\item{master}{character with full path name of a \emph{.tif} file whose extend and \code{\link[raster]{crs}}
are used to rasterize \code{sta} output. See \code{Details}}
}
\value{
A list containing a bunch of elements
}
\description{
Statistical Seasonal Trend Analysis
}
\details{
When the input is a \code{\link[base]{matrix}}, its first two columns must correspond 
to geographic coordinates. For instance, the matrix resulting from applying \code{\link[raster]{rasterToPoints}} 
to a \emph{RasterStack} follows this rule. 

\code{freq} must be either 12 (monthly observations), 23 (Landsat annual scale) or 36 (10-day composite)
as this version implements STA for time series with these frequencies.

This version sets \code{intraAnnualPeriod} to either the \code{wetSeason} or the \code{drySeason} 
of Mexico. Empirical evidence suggests that while wet season runs from May to October, dry season
runs from November to April. Should a desired STA require specific months/days, these must be
provided through \code{adhocPeriod}.

When \code{interAnnualPeriod} is not specified and \code{class(data)=numeric},
\code{interAnnualPeriod = 1:(length(data)/freq)}; when \code{class(data)} is either \code{RasterStack} 
or \code{matrix}, \code{interAnnualPeriod = 1:((ncol(data)-2)/freq)}.

Since \code{adhocPeriod} defines an inter annual period "ad-hoc", the specific days of this ad-hoc
season must be known in advance and consequently, the specific time-points (with respect to the 
time series under consideration) must be provided in a numeric vector.

When \code{plot=T} and \code{class(data)=numeric}, a generic plot is displayed. The plot area is
divided in four rectangles of equal dimensions using \code{\link[graphics]{layout}}. The rectangle in
the top displays the original time series (gray points) along with the harmonic regression fit (black line)
based on the observations indicated by \code{interAnnualPeriod}, also there are highlighted those
points (red) given by \code{intraAnnualPeriod} or \code{adhocPeriod} which are used to calculate
basic statistics. 
In the rectangle below
the output of the statistical trend analysis (performed with \code{\link[trend]{sens.slope}}) on the 
shape parameter mean is displayed; for this analysis only the years given by \code{interAnnualPeriod} 
are considered; a p-value smaller than \code{significance} is highlighted by displaying this value in red.
In the following two rectangles, similar STA (for annual and semi-annual shape parameters) are displayed.

When \code{plot=T} and \code{class(data)} is RasterStack or matrix, a \code{\link[mapview]{mapview}}
object is plotted. This mapview object contains the slope analysis maps of the mean, annual and semi-annual
shape parameters. 

When \code{save=T}, a valid \code{dirToSaveSTA} must be provided, that is, this folder should have been
created previously. In this case, the STA output is saved in \code{dirToSaveSTA}. This version
saves arrays of STA of the mean, annual and semi-annual parameters (along with their corresponding basic statistics)
in the file \code{sta_matrix_output.RData} inside \code{dirToSaveSTA}. Also, in the same directory,
the file \code{sta_progress.txt} records the progress of the STA process.

\code{master} provides the filename (with complete path) of an image free of missing values (\code{NA}).
The extent of this image must coincide with that of the RasterStack under analysis.
}
\note{
\code{save=T}, \code{dirToSaveSTA}, \code{numCores} and \code{master} are required when \code{data} is either a
RasterStack or a matrix. The aforementioned \emph{basic statistics} are: mean and standard deviation
of the time series of annual maximum and minimum as well as the global minima and maxima.
}
\examples{
# Use of interAnnualPeriod
sta_wetSeason_21016 <- sta(data = wetlands, freq = 36, interAnnualPeriod = c(2, 10, 16), 
                           plot = TRUE)
str(sta_wetSeason_21016)

# Use of adhocPeriod; include all the observations for calculating basic statistics
adhoc <- list()
beginPeriod <- (1:19-1) * 36
endPeriod <- 1:19 * 36 # adhoc$partial length must be even
adhoc$partial <- c( sapply(1:length(beginPeriod), function(s) c(beginPeriod[s]+1, endPeriod[s]) ) )
adhoc$full <- c( sapply(1:length(beginPeriod), function(s) (beginPeriod[s]+1):endPeriod[s]) )
sta(data = wetlands, freq = 36, adhocPeriod = adhoc, plot = TRUE, significance = 0.05)

}
\references{
Eastman, R., Sangermano, F., Ghimine, B., Zhu, H., Chen, H., Neeti, N., Cai, Y., Machado E., Crema, S. (2009).
\emph{Seasonal trend analysis of image time series},
International Journal of Remote Sensing, \bold{30(10)}, 2721--2726.
}
