% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper_vis.R
\name{stat_summery}
\alias{stat_summery}
\title{Create statistics summary of nodes}
\usage{
stat_summery(obj_mapper, groups_ind, dat = NULL, folder = "",
  add_surv_analysis = FALSE, add_analysis_java = NULL, color_code)
}
\arguments{
\item{obj_mapper}{An object of class \code{TDAmapper}.}

\item{groups_ind}{A vector of group names each of the samples belongs to.}

\item{dat}{Optional. A dataframe contains additional information of samples.
If \code{add_surv_analysis==TRUE}, \code{dat} must be provided with columns
\code{age_at_diagnosis} and \code{days_to_death} describing the time at
diagnosis and time to death at the time of observation.}

\item{folder}{The name of the folder to save the generated networks.}

\item{add_surv_analysis}{A logical object. \code{TRUE} if median time at diagnosis
and median time to death should be included.}

\item{add_analysis_java}{A vector of Javascript codes summerizing customized
statistics.}

\item{color_code}{A color code dataframe.}
}
\value{
A list includes the statistics and pie plots in folder
  \code{/figures}. Details of elements in the list can be found in Detail.
}
\description{
\code{stat_summery} returns statistics summaries of nodes as well as a vector
of Javascript codes describing the summaries.
}
\details{
This function summarizes the nodes returned by
\code{\link[TDAmapper]{mapper}}. Basic statistics included in this function
are the number of samples \code{N}, the percentage of samples from different
groups \code{groups_per_node}, the groups dominate the nodes
\code{dominant_group}, the percentage of samples from the dominated group
\code{dominant_percent}. It also generates pie plots for nodes which are
saved in \code{/figures}.

If additional information is provided, the function can provide further
statistics including median time at diagnosis and median time to death.

The users can also define and add customized statistics to this function. The
customized statistics should be wraped into a vector of Javascript codes for
each of the nodes and passed to \code{add_analysis_java} following the same
order of nodes as in \code{obj_mapper}.
}
\examples{

}
