% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlands-data.R
\docType{data}
\name{wetlands}
\alias{wetlands}
\title{Wetland Time Series}
\format{An object of class \code{"numeric"}.}
\usage{
data(wetlands)
}
\description{
Time series from a RasterStack of moisture index recorded on Mexican wetlands
}
\keyword{datasets}
