% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{chicken_generator}
\alias{chicken_generator}
\title{Chicken foot simulation data generator}
\usage{
chicken_generator(seed = 1)
}
\arguments{
\item{seed}{A scaler determining the seed to be used.}
}
\value{
A dataframe of the simulated data. The first column denotes the
  groups of the points and the last three columns are the coordinates of data
  points.
}
\description{
\code{chicken_generator} generates chicken foot simulation data.
}
\details{
\code{chicken_generator} generates chicken foot simulation data with 450 data
points in total. It consists of three toes and one shank. Each of the toes
contains 105 data points and the shank contains 135 data points.
}
\examples{
chicken_generator(1)

}
