% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta.R
\name{plot.staMatrix}
\alias{plot.staMatrix}
\title{Plot method for \code{sta} function}
\usage{
\method{plot}{staMatrix}(x, significance = NULL, master, ...)
}
\arguments{
\item{x}{an object of class "staMatrix"}

\item{significance}{numeric indicating significance of each shape parameter trend}

\item{master}{RasterLayer used to transfer STA output to raster layers.}

\item{...}{additional plot parameters}
}
\description{
This function displays some maps of \code{\link[mapview]{mapview-class}}
}
\seealso{
\code{\link[sta]{sta}}, \code{\link[sta]{getMaster}}, \code{\link[geoTS]{matrixToRaster}}
}
